added heavy refactor
